requirements.txt
streamlit
requests
beautifulsoup4
pandas
apscheduler
python-dotenv
openai
google-generativeai
sqlalchemy
typing_extensions
